pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lightberry-ai"
version = "0.1.0"
description = "LiveKit Audio Streaming with Tool Execution"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Lightberry AI", email = "info@lightberry.ai"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers", 
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Communications :: Conferencing",
]
keywords = ["livekit", "audio", "streaming", "real-time", "tools", "ai"]
requires-python = ">=3.8"
dependencies = [
    "livekit>=0.10.0",
    "livekit-api>=0.5.0", 
    "sounddevice>=0.4.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.21.0",
    "aiohttp>=3.8.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0"
]

[project.urls]
Homepage = "https://lightberry.ai"
Repository = "https://github.com/lightberry-ai/lightberry-python-sdk"
Documentation = "https://docs.lightberry.ai"
"Bug Tracker" = "https://github.com/lightberry-ai/lightberry-python-sdk/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["lightberry_ai*"]

[tool.setuptools.package-data]
lightberry_ai = ["py.typed"]